[BC] Fixing dtype

+++ SciDataTool/Functions/interpolations.py
@@ -145,7 +145,7 @@ def get_interpolation_step(values, axis_values, new_axis_values, index):
         else:
             new_shape = list(values.shape)
             new_shape[index] = len(new_axis_values)
-            new_values = zeros(tuple(new_shape))
+            new_values = zeros(tuple(new_shape), dtype=values.dtype)
             for i in range(len(new_axis_values)):
                 for j in range(len(axis_values) - 1):
                     if (
